Adding Profile = srxgui
To load into ipython use the command:
ipython --profile=srxgui
